Update home page
Use markdown. Homepage now reflects the new publisher-friendly angle. In other words the primary goal of Jekyll-bootstrap is to get the user up and running with their own blog as fast as possible. To do this we employ 'jekyll-bootstrap-core' as a fresh install framework for users to use as the base of their blog