added pagination proof of concept
- Changed the set theme for existing CSS compatibility
- added paginate variable
- added added code to run locally
- changed index to html in order to use helpers
- renamed index.md to intro.md 
not already in the theme
- added posterator helper
	- based off work by nhoss2
	- include the helper on any html page
	- include 'is_index: true: in the meta data

I chose the name posterator instead of paginator 
because the name paginate is already in use.  

P.S. I don't know what I am doing.  Please be on 
the look out for mistakes.  I have it working 
locally to some degree, but it has not been 
extensively tested.
Signed-off-by: Bret Comnes <bcomnes@gmail.com>