Use rake task to swtich themes
the theme switcher file is too complicated. Ideally we'll get the ability to pass in liquid variables to the include tag, but untill then this is the best implemention I've got. The strategy being simply redefining references within the layout files. This way all themes and theme updates are silo'ed in their own directories.